PHPLIB-1386 Fix deprecation: use E_USER_DEPRECATED, version 2.4, validate coll field
