+++ src/Builder/Stage.php
@@ -16,7 +16,6 @@ use MongoDB\Builder\Type\Optional;
 use MongoDB\Builder\Type\QueryInterface;
 use stdClass;
 
-use function assert;
 use function get_debug_type;
 use function is_array;
 use function is_string;
@@ -48,7 +47,7 @@ final class Stage
      * Writes the resulting documents of the aggregation pipeline to a collection. To use the $out stage, it must be the last stage in the pipeline.
      *
      * @see https://www.mongodb.com/docs/manual/reference/operator/aggregation/out/
-     * @param Document|Serializable|array|stdClass|string   $coll       The output collection name.
+     * @param Document|Serializable|array|stdClass|string   $coll       The output collection name. Passing a non-string value is deprecated since 2.4.
      * @param Optional|string                               $db         The output database name. If omitted, defaults to the current database.
      * @param Optional|Document|Serializable|array|stdClass $timeseries Specifies the configuration to use when writing to a time series collection.
      * The timeField is required. All other fields are optional.
@@ -61,7 +60,7 @@ final class Stage
         Optional|Document|Serializable|stdClass|array $timeseries = Optional::Undefined,
     ): OutStage {
         if (! is_string($coll)) {
-            trigger_error(sprintf('Since mongodb/mongodb 2.3: %s::%s() first argument must be the collection name. Passing "%s" is deprecated', self::class, __FUNCTION__, get_debug_type($coll)), E_USER_DEPRECATED);
+            trigger_error(sprintf('Since mongodb/mongodb 2.4: %s::%s() first argument must be the collection name. Passing "%s" is deprecated', self::class, __FUNCTION__, get_debug_type($coll)), E_USER_DEPRECATED);
 
             if ($db !== Optional::Undefined || $timeseries !== Optional::Undefined) {
                 throw new InvalidArgumentException('When passing a non-string first argument, the $db and $timeseries arguments must not be passed.');
@@ -73,13 +72,15 @@ final class Stage
                 $coll = (array) $coll;
             }
 
-            assert(is_array($coll));
+            if (! is_array($coll) || ! isset($coll['coll']) || ! is_string($coll['coll'])) {
+                throw new InvalidArgumentException('When passing a non-string first argument, it must be an array or object with a "coll" string field.');
+            }
 
             if (isset($coll['db'])) {
                 $db = (string) $coll['db'];
             }
 
-            $coll = (string) $coll['coll'];
+            $coll = $coll['coll'];
         }
 
         return new OutStage($coll, $db, $timeseries);
